backend/requirements.txt
django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
prisma==0.13.1
pyjwt>=2.8.0
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
psycopg[binary]==3.2.9
python-dateutil==2.9.0.post0
